Server/requirements.txt
APScheduler==3.10.1
Django==4.1.7
django_apscheduler==0.6.2
django_filter==23.1
djangorestframework==3.12.0
djangorestframework_simplejwt==5.2.2
docker==6.0.1
mysqlclient==2.1.1
drf_yasg2==1.19.4